Full fix for toolbar-add tests
This is deeply unsatisfying, but this test will fail if it is not run
first. It appears that if the editor has previously been in the DOM,
then there are two renders of the toolbar during the test: one based on
the previous coordinates and then one with the new coordinates on the
right.

A `setTimeout(..., 0)` gives the component enough time to draw in the
right place.