Fix problems with this.render returning `null`
This takes the pattern rolled out in the last few commits and migrates
it to an abstraction where we return a ref instead of the result of
`ReactDOM.render` in out `this.render` testing utility. This will
probably fix a few failures, and I verified that the tests that I
previously fixed stay fixed with this change in place.