Make richcombo-bridge tests resilient against changes in order
I fixed these tests in 73acf3fe, or at least I thought I did. Turns out
that last one only passed because I was testing it with `it.only`; if
you run it after the others, the click listener does not get called.

The reason why is as follows:

- Out test_richcombobridge/plugin.js is a test-only plugin that adds a
  new rich combo with a hard-coded name. Even though we end up adding a
  new rich combo in every `beforeEach` block, we end up using the same
  name.
- In the richcombo bridge itself (src/components/uibridge/richcombo.jsx)
  we have stash all these combos in an internal cache,
  `RICH_COMBO_DEFS`, indexed by name. That means that only the first def
  ends up getting stored and never overwritten.
- So, even though we have a unique click handler for each test, only the
  first-registered one can ever be called. This is because the
  component's `_onClick` implementation will look up the combo in the
  cache and get the first one and call that one's click handler.
- You can see this if you add log statements which log the editor id at
  each step; you'll see something like this:

    1. Register combo for editor id cke_2; it gets cached.
    2. ... run various tests.
    3. Register combo for editor id cke_42; it doesn't get cached
       because we already have a combo cached for name
       "ButtonRichCombo".
    4. Fire click on editor id cke_42; see the handler from id cke_2 get
       called.

So, the fix is to register a new combo for each test block, which we
achieve by appending an incrementing counter to each name.

Is this a problem outside the test suite? I don't know. One mitigating
factor is that the cache keys comprise not only the rich combo name, but
the editor name (eg. "editable", at least in the test suite; presumably
people use different names in prod, especially when there are multiple
instances on a page) as well.