Fix button-bridge tests
These were broken for much the same reason as the richcombo-bridge
tests were and which I fixed in the previous commit.

Just like richcombo-bridge, button-bridge maintains an internal cache,
this time called `BUTTON_DEFS`, which cases definitions by name and our
use of a hardcoded name here ('ButtonClick') meant that only our
first-registered click handler would ever receive events.

As before, the fix here is to generate a unique name for each definition
in the test.