Fix button-link-edit-browse.jsx tests
I fixed these tests the other day in 40eef9d8b43fa7ac7dfd39135e642340
but now they are broken again, for a different reason:

If you look in the implementation for where the click is handled, you
see this:

```
_browseClick = () => {
        const editor = this.context.editor.get('nativeEditor');
        const url = editor.config.documentBrowseLinkUrl;
        const browseLinkCallback = editor.config.documentBrowseLinkCallback;
        const linkTarget = this.linkEditButtonRef.current
                ? this.linkEditButtonRef.current.state.linkTarget
                : '';

        const changeLinkCallback = selectedItem => {
                this._updateLink(selectedItem.value, linkTarget, selectedItem.href);
        };

        if (Lang.isFunction(browseLinkCallback)) {
                browseLinkCallback.apply(null, [editor, url, changeLinkCallback]);
        }
};
```

The link callback is clearly not being called with `editor, url, target`
as is being asserted, so I am not sure how this was passing previously.
It is in fact being called with `editor, url, callback`, so I updated
the test to reflect that.

Looking at the history for the component to see when the behavior might
have changed, the last commit is f5e4146349707b5537487f94e62b742d, but
that doesn't change the signature of the callback. Before that was
9e56b2bbe81074e5fcd0c4bbef28a3c086e, but that didn't change the
signature either. Looking still further back we see
387c01ac6006d3a307d41ce5ab7a5ddb is the one that changed the signature,
so the changes in this commit are correct. I have no idea why the tests
seemed to be passing before... I probably just made a mistake in
thinking they were green.