Silence React warnings in button-link-target-edit tests
Fix this warning:

```
Warning: `value` prop on `input` should not be null. Consider using an
empty string to clear the component or `undefined` for uncontrolled
components.
```

Our ternaries in the component were intended to ensure that we never set
properties to `null`, but in the artificial environment of the test
suite they weren't doing that.

`_getInitialState` creates a `new CKEDITOR.Link` from the selection, but
there is no selection, so `link` is created but `link.getAttribute(...)`
always returns `null`.

Solution here is to tweak the ternaries to make sure we always fall back
and never set these values to `null`.