Be more resilient in post-test cleanup
While investigating a failing test I sometimes want to comment out or
change code in order to test out a theory. For example, I might want to
skip the `beforeEach` hook in some cases. However, if I do that, then I
have to skip the `afterEach` hook too otherwise it will blow up trying
to access `this._editable` and `this.container`.

Make it possible to do this kind of thing without throwing an error by
adding a couple of `if` guards.