Fix embed.js tests
Alas, I have to reintroduce the hack that I removed in
c7f77084e7a5631e4814c92bf1. Notice though that I am not applying it in a
blanket fashion everywhere -- that just makes other failures pop up
elsewhere -- but instead am targeting it to the two places that need it.
There are only two other failing tests in Firefox visible with `npm run
test:debug` right now. If they require a similar fix I'll extract out
the hack.

This is some really baffling behavior. To summarize:

- Tests only fail on Firefox, and only if a test has run before.
- In fact, it is not even necessary to run a test before to trigger the
  failure: just adding a "div" with a "contenteditable" attribute to the
  DOM is enough (removing it makes no difference).
- As I was single-stepping through the debugger, I saw CKEDITOR
  incorrectly attaching itself to the first contenteditable that it
  found, even if it did not have the "#editable" ID that we are passing
  in out call to `CKEDITOR.inline`, and even if we pass the actual DOM
  object instead of passing the ID name.
- When you set the selection using this hack, the problem magically goes
  away.

I'd still like to try this with a non-debug build of CKEDITOR
-- see https://github.com/liferay/alloy-editor/issues/1150 -- so I can
see what the actual error that is being thrown is, but for now it's just
an inscrutable 'TypeError: "B is undefined"'.