Automate CKEditor build process
This commit adds a script for performing a build, invoked like this:

    scripts/build/build-ckeditor.sh 4.11.2

Update CONTRIBUTING.md to reflect the existence of the tool, and made
sure that dirty status in the ckeditor-dev submodule won't clutter up
`git status` output (it's normal to dirty the submodule because we copy
our own "build-config.js" into it).

Test plan: run the script and see that the build worked and the only
changes made to the files were timestamp updates.

Closes: https://github.com/liferay/alloy-editor/issues/1124