Turn off lint rules that we get via extension
We don't need either of these rules any more because they are bundled
with the new version of eslint-config-liferay.

Test plan: Edit a file to introduce problems (eg. a `for`/`of` loop, and
a function that has unused arguments), and see lints show up when I run
`npm run lint:quiet`.