feat: Use "yarn" instead of "npm" (#1143)
Moving to yarn is good for consistency with our other projects (we are
using it in DXP now, and also in a number of our smaller GitHub
projects), and it brings us some neat features like `yarn
upgrade-interactive`.

It also gives us a more compact lockfile that is less likely to produce
merge conflicts:

7 files changed, 8783 insertions(+), 13324 deletions(-)

Test plan:

`yarn install` and make sure all the scripts still work:

- `yarn build:assets`
- `yarn build`
- `yarn dev`
- `yarn format`
- `yarn format:changed`
- `yarn format:check`
- `yarn lint`
- `yarn lint:changed`
- `yarn lint:fix`
- `yarn lint:quiet`
- `yarn prepublishOnly`
- `yarn start` (and visit demo at http://0.0.0.0:9000/)
- `yarn test`
- `yarn test:debug` (and verify in browser that everything is still
  green)

Not tested: `yarn testSaucelabs` (CI will do that).

Closes: https://github.com/liferay/alloy-editor/issues/1143