fix: missing PropTypes on AlloyEditor.React (#1200)
We still use PropTypes internally, even though it is no longer
bundled with React. Just had a report from @ricardocousolr about
a failure on current master stemming from trying to access
`React.AlloyEditor.PropTypes`. Given that we still use it, we should
expose it in the same way as before as people may be relying on it.

Test plan: swap out the build in the demo for the `no-ckeditor` one
(which exposes React, and which we use in DXP), and in browser console
verify that I can access `AlloyEditor.React.PropTypes`.

```
diff --git a/scripts/build/gulp/demo/index.html b/scripts/build/gulp/demo/index.html
index 93f8964c..62fb48e3 100644
--- a/scripts/build/gulp/demo/index.html
+++ b/scripts/build/gulp/demo/index.html
@@ -10,7 +10,8 @@

         <link href="alloy-editor/assets/alloy-editor-ocean.css" rel="stylesheet">

-        <script src="alloy-editor/alloy-editor-all.js"></script>
+        <script src="https://cdn.ckeditor.com/4.11.3/standard/ckeditor.js"></script>
+        <script src="alloy-editor/alloy-editor-no-ckeditor.js"></script>

         <style>
             body {
```

Closes: https://github.com/liferay/alloy-editor/issues/1200