feat: Add deploy.sh script to publish to GitHub pages (#1212)
This commit adds a shell script that takes the contents of the "public/"
directory and creates a new commit from them on the "gh-pages" branch.

If you run `./deploy.sh`, you should see something like the following
(the dots are progress indicators, printed one per file as we go seeing
as we are forking a lot and it isn't super fast; it takes about 11
seconds on my machine to run the script):

```
Working from top level: /Users/greghurrell/code/alloy-editor/website
Clearing index
Adding public/ files to index
.......................................................................
.......................................................................
.......................................................................
.......................................................................
.......................................................................
.......................................................................
...................................
Created tree object 462d4368bb55c356147afe3c4e001ce2434dbdb3
Creating commit object on gh-pages branch
commit 6f4e2d29ae46ef339571877f7e42435883a079e5 (gh-pages)
Author: Greg Hurrell <greg.hurrell@liferay.com>
Date:   Mon Apr 1 18:38:44 2019 +0200

    Autogenerated public/ content from f606edab
Restoring index to prior state
```

Caveats: because `public/` is in .gitignore, we can't produce this from
the `HEAD` and instead just take whatever is on the disk. Hopefully that
corresponds to what you'd get from running `yarn build` on HEAD, but
there are no guarantees.

Likewise, it only creates the commit on the "gh-pages" branch -- it
doesn't actually push it -- so that you can review it.

Test plan: run the command as above and see it produce this commit on
the branch:

https://github.com/liferay/alloy-editor/commit/6f4e2d29ae46ef339571877f7e42435883a079e5

Related: https://github.com/liferay/alloy-editor/issues/1212