chore: Update to latest clay-css (#1220)
We're using clay-css for the spritesheet, but we had it pinned to an old
version. In liferay-portal, we're using clay-css 2.11.1 right now, and
there's no reason why we shouldn't just use the same here. So I'm
specifying a looser range. Note that the latest is actually 2.12.0, but
I only ask for "^2.11.1" because I want `yarn` in the portal to resolve
clay-css to a single version.

Test plan: `yarn dev && yarn start`

Closes: https://github.com/liferay/alloy-editor/issues/1220