fix: console errors in Windows-based build (#1228)
It turns out that the "test" property in the Webpack build needs to be
in a different format on Windows, otherwise it never matches, which
means that our file at "scripts/build/version.js" doesn't pass through
the val-loader, and in turn means that we produce untransformed output
that IE 11 will choke on (an arrow function, and even if we fixed it, it
would probably choke on the backticks).

You can make the build work on Windows by passing
`path.resolve('scripts/build/version.js')` instead; this produces an
absolute path with the appropriate Windows-style backslashes. But then
the build doesn't work on macOS. If you try to make it a relative path
(`path.relative('', path.resolve(...))`) then Webpack rejects it.

So, the solution is to make a RegExp with the backslashes (if present)
escaped.

Test plan: build on macOS and Windows (with `yarn build`, `yarn start`)
and confirm console errors are gone and version is correctly transpiled.
Note that the demo misbehaves still on IE 11 (not console errors, but
some misbehavior of the editor itself, but I think that's expected).

Closes: https://github.com/liferay/alloy-editor/issues/1228