chore: make Travis use Node v11 (1.x) (#1237)
We needed to force the use of v11 on the master branch (see #1238)
because v12 was causing CI builds to blow up. This is not a problem on
the 1.x branch because we were still using v8. But we should use the
same version on both branches for consistency.

Related: https://github.com/liferay/alloy-editor/issues/1237