feat: check that dist file contains expected version number (#1257)
Add a script that checks the main built dist file to see whether it
contains the right version number and use it in "prepublishOnly", which
runs "Run BEFORE the package is prepared and packed, ONLY on npm
publish":

https://docs.npmjs.com/misc/scripts

Could also have just re-run the build here, but that is slow, and we're
already running it once on "yarn version".

Test plan: Go through pretend release process and see:

```
$ node scripts/publish/check.js
🔍 Checking alloy-editor-all.js for presence of version identifier: 2.0.3
✅ Found version 2.0.3 in alloy-editor-all.js
```

Also, purposely break dist file in various ways (remove the file, remove
the sentinel, remove the version string etc) and see errors.

Closes: https://github.com/liferay/alloy-editor/issues/1257