refactor: do not commit dist files (#1300)
Committing build artifacts is an anti-pattern, for the most part, and we
stopped doing that on the master branch a while back (in 1911b8756eec,
"Don't store dist/* contents in the repo", 22 January 2019). So let's
apply the same change here on the 1.x branch.

Doing this now because we found the 1.5.25 build was broken (had some
missing files in the release tarball) and this is a step towards making
our builds more deterministic.

See: https://github.com/liferay/alloy-editor/issues/1300