refactor!: don't run "clean" tasks as part of build tasks (#1300)
Cleaning and building are distinct operations, conceptually, so it is
wrong to fuse them inseparably together. Instead, you should be able to
do any of the following:

- clean; or
- build; or
- clean and build

So, I removed the auto-clean dependencies from the build tasks, and
additionally removed the "clean-fonts" and "clean-api" tasks that are
already covered by "clean-dist".

"clean-dist" itself now gets renamed to just "clean" because it is the
only cleaning task.

See: https://github.com/liferay/alloy-editor/issues/1300