refactor: add missing "files" property to package.json
This is similar to the change that we made on the master branch in
6582d82df507bab41903d2 ("Switch package.json to whitelist included
files", 22 January 2019), but with a few extra tweaks as noted below.

In testing my recent changes with `npm publish --dry-run`, I see it
including an enormous amount of garbage in the tarball that has no
business in being there:

    npm notice === Tarball Details ===
    npm notice name:          alloyeditor
    npm notice version:       1.5.25
    npm notice package size:  34.6 MB
    npm notice unpacked size: 137.7 MB
    npm notice shasum:        bfbaa165ec06c215729734f0ab7e3144875bf06a
    npm notice integrity:     sha512-VPPNZjj0JmcDg[...]ZHWAWcxOVkJLQ==
    npm notice total files:   11098

Note that because "dist" is gitgnored, NPM will not include it despite
the explicit "files" directive, unless we add an .npmignore that
overrides the .gitignore.

I deleted the now-unnecessary top-level .npmignore (it was unnecessary
even before the "files" change in any case, because all the stuff in it
was also in the .gitignore file).

Docs on how .npmignore and .gitignore interact are here:

https://docs.npmjs.com/misc/developers#keeping-files-out-of-your-package