refactor: clean up "post-cleanup" function
Label the different groups here as the reason why we want to remove them
may be quite mysterious (and arguably, we shouldn't be dumping temporary
files in the "dist" folder in any case, but that would be a bigger fix
than I want to take on at this point).

And clean up the intermediate "assets/css" files that were being
included in the dist package. We discovered those via manual inspection
of the test release that I prepared in the last commit.

What was in v1.5.24:

    https://gist.github.com/wincent/aa08eda2e189e97646c54f2c47557fb2

What was in the test release:

    https://gist.github.com/wincent/a25a142c9820b45da08c54941ee41f0e