refactor: remove unnecessary length guards
If `heightStyles` and `widthStyles` are truthy, it is because they are
non-`null` arrays. On these arrays, `length` is always guaranteed to
be at least 1, because at the very least, the full matching string is
always returned as the 0th element (and in the case of these regexps,
we'll always have a 1st element too, due to the use of a capturing
group).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/exec