feat: export 2.x HOCs and 1.x mixins (#1314)
Exposes:

- The 2.x HOCs as `AlloyEditor.Base.ButtonCommand` (etc) so that people
  can write custom buttons using the new ES6-class based style.
- The 1.x mixins as `AlloyEditor.Compat.ButtonCommand` (etc), for people who
  have custom buttons written using the old (`React.createClass`) style.

The obvious downside here is that we're shipping more code now,
increasing the bundle size and overall complexity, but we think it's
worth it because we are prepared to pay quite a high cost in the name of
maintaining backwards compatibility.

Having said that, we are not going to go out of our way to document the
`Compat` mixins at this time, so as not to overwhelm users; this is just
something we want to have up our sleeve to be able to offer it if the
need arises (and it may not arise, given that custom buttons probably
aren't often implemented).

NOTE: to make the overall PR readable this commit just copies the files
into place without any significant changes. Subsequent commits will
apply formatting, fix lints, and actually set up the default exports in
the compat files.

Test plan: will check build and demo works; not doing a full integration
test with custom buttons at this time.

Closes: https://github.com/liferay/alloy-editor/issues/1314