chore: quieten some of the noiser (JSDoc-related) lints (#1314)
This drops us from 1022 warnings to 648. We think this is ok because:

- This project is mostly in "maintenance mode" at this point.
- Realistically, these warning are never going to be addressed.
- If we were to invest in this project in the future, we believe
  TypeScript is a better investment than JSDoc for this kind of
  internal-only documentation, because it provides us with no only
  information about the API, but enforcement of correct usage.

A large number of the remaining warnings come from 'sort-keys', which we
may or may not address in the future (the reason we haven't done so so
far is that it is very manual and error-prone, so we don't want to incur
an unnecessary risk, especially when the cost is so high and the benefit
so marginal).