fix: avoid JS error computing selection position
Noticed while testing https://github.com/liferay/alloy-editor/pull/1414
a bunch of console spew coming from the `getClientRect()` calls in
selection-position.js. These prevent the toolbar from being shown. I
gather I am not the only one who sees this, because in @diegonvs's
screencap gif you can also see there is no toolbar.

This looks like a regression introduced in bef3ce28. I don't know why we
didn't catch it during initial testing; that made me think that it might
actually be the subsequent CKEditor update that exposed the problem, but
no, when I go back to bef3ce28 I can repro the error.

The problem is that `startContainer` may be a text node, which doesn't
have a `getClientRect()` method on it:

https://ckeditor.com/docs/ckeditor4/latest/api/CKEDITOR_dom_text.html

So, this commit tries to get the parent element in that case, and if we
have one, we proceed; elements do have the `getClientRect()` method:

https://ckeditor.com/docs/ckeditor4/latest/api/CKEDITOR_dom_element.html

Test plan:

Run demo (`yarn start`) and see that I can click in a cell to see the
table-specific toolbar; I can select text in a cell to see text-specific
toolbar.

Also run build and test it out in liferay-portal:

    yarn build

    # Use zsh trick...
    setopt extendedglob

    # ...to copy everything but node_modules and website:
    cp -R ^(node_modules|website) ~modules/node_modules/alloyeditor

    # ... cd to path frontend-editor-alloyeditor-web
    j alloy

    # ... deploy
    gradlew deploy -a

On master, I don't see the console spew or the missing toolbar (for some
unknown reason).  With this PR, toolbar continues to appear correctly,
so whatever we're doing here is compatible with the CKEditor version in
liferay-portal.

But note that there is some very funky stuff going in master at
the moment, where you can't click to move the cursor (you have to
use the keyboard. No JS errors in the console. I am aware of some
CKEditor-related work that is going on though that may already include
an inbound fix for that.