docs: fix version numbers in changelog
I was meaning to cut a patch release when I generated the changelog, but
then I ran `yarn version --minor` instead of `yarn version --patch` by
mistake. If this were an unintended major bump I would be pretty
concerned, but it is only a minor bump, so I feel a bit better about
that.