chore: update to latest version of @liferay/eslint-config
Because, as noted in:

    https://github.com/liferay/liferay-frontend-projects/pull/429

We will be turning off the `no-for-of-loops` rule by default, which
means we have to turn it back on in places that might still want it,
like this repo.

However, while doing this, I figured I might as well update to a newer
version of `eslint-config-liferay` in this repo (really, that means
moving to `@liferay/eslint-config`). Turns out that this, and updating
ESLint like I am here means that we are going to need this PR too:

    https://github.com/liferay/liferay-frontend-projects/pull/430

because the way we monkey-patch ESLint needs to be updated.

Note the `typescript` dependency is added just because some of our rules
depend on it, and we happen to use it everywhere else anyway (and back
in the day, I didn't want to add a hard-dependency on it to our ESLint
config... this is something for us to reevaluate though, probably).

Next commit will apply autofixes for new rules. And after that I'll turn
on the `no-for-of-loops` rule which is the entire reason I am doing
this.