fix: reconcile Prettier and ESLint
They disagree about where comments should go. As this is a legacy
project, I am taking the easy way out and just turning off the
problematic rule.

The "right" fix would be doing formatting with our Prettier wrapper from
`@liferay/npm-scripts`, but we can't easily `yarn add` that here because
the use of ancient dependencies in this project leads us to have an
irreconcilable dependency on "human-signals" 2.1.0:

  error human-signals@2.1.0: The engine "node" is incompatible with this module. Expected version ">=10.17.0". Got "10.15.1"

Contrast with liferay-frontend-projects, where we are using v1.1.1.